provided common layout for most pages
